test #4 :  사용자 입력 기능 테스트 작성
